Finish experiment
