+++ modules/pun/Cargo.toml
@@ -4,10 +4,10 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0.66"
 serde = { version = "1.0.147", features = ["derive"] }
 serde_json = "1.0.87"
 
+libcommand.workspace = true
 wit-log.workspace = true
 wit-sync-request.workspace = true
